Fixed NPE during configuration: Integer.getInteger() does not parse ints
Fixed a NullPointerException during the initialization of the connector:

Integer.getInteger() and Long.getLong() do not parse numbers. They fetch
system-properties by the given name. Hence, they alway returned null, so
that the connector could not be configured.

I wonder, if that has ever worked...

Nevertheless, with this fix the connector works correct (at first glance).